+++ packages/stream_chat/lib/src/core/api/requests.dart
@@ -176,8 +176,8 @@ class ThreadOptions extends Equatable {
   const ThreadOptions({
     this.watch = true,
     this.replyLimit = 2,
-    this.participantLimit = 100,
-    this.memberLimit = 100,
+    this.participantLimit = 10,
+    this.memberLimit = 10,
   });
 
   /// If true, the client will watch for changes in the thread.
@@ -192,12 +192,12 @@ class ThreadOptions extends Equatable {
 
   /// The number of thread participants to return per thread.
   ///
-  /// Defaults to 100.
+  /// Defaults to 10.
   final int participantLimit;
 
   /// The number of members to return per thread.
   ///
-  /// Defaults to 100.
+  /// Defaults to 10.
   final int memberLimit;
 
   /// Serialize model to json